Support some jp.nephy.jsonkt.delegation API
