+++ src/commonMain/kotlin/jp/nephy/jsonkt/Reflection.kt
@@ -1,0 +1,5 @@
+package jp.nephy.jsonkt
+
+import kotlin.reflect.KProperty
+
+expect val KProperty<*>.returnsNullable: Boolean
