+++ src/commonMain/kotlin/jp/nephy/jsonkt/delegation/CharProperty.kt
