+++ src/commonMain/kotlin/jp/nephy/jsonkt/delegation/DoubleProperty.kt
