+++ src/commonMain/kotlin/jp/nephy/jsonkt/delegation/EnumProperty.kt
