+++ src/commonMain/kotlin/jp/nephy/jsonkt/delegation/StringProperty.kt
