+++ src/jsMain/kotlin/jp/nephy/jsonkt/ReflectionJS.kt
@@ -1,0 +1,8 @@
+package jp.nephy.jsonkt
+
+import kotlin.reflect.KProperty
+
+actual val KProperty<*>.returnsNullable: Boolean
+    // TODO: Wait for Reflection API implementation
+    //         `returnType.isMarkedNullable`
+    get() = true
